+++ include/unity/scopes/qt/internal/QUtils.h
@@ -18,13 +18,15 @@
 
 #pragma once
 
+#ifndef _ENABLE_QT_EXPERIMENTAL_
+#error You should define _ENABLE_QT_EXPERIMENTAL_ in order to use this experimental header file.
+#endif
+
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wctor-dtor-privacy"
 #include <QtCore/QVariant>
 #pragma GCC diagnostic pop
 #include <unity/scopes/Variant.h>
-
-#define EXPORT __attribute__((visibility("default")))
 
 namespace unity
 {
@@ -35,12 +37,15 @@ namespace scopes
 namespace qt
 {
 
-///@cond
-EXPORT QVariant scopeVariantToQVariant(unity::scopes::Variant const& variant);
-EXPORT unity::scopes::Variant qVariantToScopeVariant(QVariant const& variant);
-EXPORT QVariantMap scopeVariantMapToQVariantMap(unity::scopes::VariantMap const& variant);
-EXPORT VariantMap qVariantMapToScopeVariantMap(QVariantMap const& variant);
-///@endcond
+namespace internal
+{
+
+QVariant variant_to_qvariant(unity::scopes::Variant const& variant);
+unity::scopes::Variant qvariant_to_variant(QVariant const& variant);
+QVariantMap variantmap_to_qvariantmap(unity::scopes::VariantMap const& variant);
+VariantMap qvariantmap_to_variantmap(QVariantMap const& variant);
+
+}  // namespace qt
 
 }  // namespace qt