+++ src/scopes/qt/XmlReader.cpp
@@ -17,7 +17,6 @@
  */
 
 // allow experimental headers
-#define _ENABLE_QT_EXPERIMENTAL_
 
 #include <unity/scopes/qt/XmlAsyncReader.h>
 #include <unity/scopes/qt/XmlReader.h>